Clases del 24_10

+++ SwitchDash/InterfaceModule/src/main/java/es/ucm/fdi/moviles/interfacemodule/Image.java
@@ -3,8 +3,8 @@ package es.ucm.fdi.moviles.interfacemodule;
 public interface Image
 {
     //Get image width
-    public float getWidth();
+    public int getWidth();
 
     //Get image heigth
-    public float getHeigth();
+    public int getHeigth();
 }
